+++ serdes_json/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.0
+
+* Make `dynamic` primitive;
+* Added `SerdesJsonField` annotation.
+
 ## 0.1.0+1
 
 * Fixed README.md links